requirements.txt
streamlit>=1.29.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0
langchain>=0.0.267
langchain-cerebras
python-dotenv>=1.0.0
scipy>=1.10.0
joblib>=1.2.0
pydantic>=2.0.0
requests>=2.28.0
pillow>=9.0.0
altair>=4.2.0
beautifulsoup4>=4.11.0
xgboost


